// src/types/ingredient.ts
export type TIngredientType = 'bun' | 'sauce' | 'main';

export type TIngredient = {
  _id: string;
  name: string;
  type: TIngredientType;

  proteins: number;
  fat: number;
  carbohydrates: number;
  calories: number;

  price: number;

  image: string;
  image_mobile: string;
  image_large: string;

  __v?: number;
};

export type TConstructorIngredient = TIngredient & { uuid: string };